requirements.txt
flask==3.0.0
redis==5.0.1
cryptography==41.0.7
flask-babel==4.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
